refactor: simplify multi-agent architecture with clean tool separation
Architecture Changes:
- Sentiment Agent: Pure news-based analysis using Google Search API only
- Tech Agent: Polygon.io primary (5/min) with Alpha Vantage fallback (25/day)
- Strategy Agent: Pure aggregator with no direct data access tools
- Removed market heat calculation from sentiment agent

Tool System Cleanup:
- Removed all deprecated tools (SEC, FMP, Finnhub, Yahoo Finance)
- Created Google Search simple tool for sentiment analysis
- Cleaned tools.py from 1100+ to 180 lines
- Updated base_agent.py imports to match active tools only

File Changes:
- sentiment_agent.py: Rewritten for news-only sentiment analysis
- tools.py: Comprehensive cleanup with clear agent-tool assignments
- base_agent.py: Removed deprecated imports and functions
- google_search_simple.py: New clean news tool implementation
- deprecated/: Archived old tool implementations

Issues Resolved: #170 (SPY momentum error), #172-179 (architecture simplification)
