+++ src/agents/__init__.py
@@ -3,4 +3,3 @@ from .sentiment_agent import SentimentAgent
 from .strategy_agent import StrategyAgent
 from .coordinator_agent import CoordinatorAgent
 from .tech_agent import TechAgent
-from .coordinator_agent import CoordinatorAgent
